+++ bin/possum
@@ -2,7 +2,7 @@
 
 import sys
 
-from possum import main
+from possum.cli import main
 
 
 if __name__ == '__main__':
