+++ possum/exc.py
@@ -1,0 +1,6 @@
+class PossumException(Exception):
+    """Base Possum Exception"""
+
+
+class PipenvPathNotFound(PossumException):
+    """Pipenv could not be located"""
